메뉴얼 및 업데이트 로그 작성. id 자료형 변경(ulong->int)

+++ project/Capstone_Design/Assets/BaseSystem/PlayInfo.cs
@@ -10,7 +10,7 @@ public class PlayInfo
     int value_int;
     float value_float;
     string value_string;
-    public ulong id { get; }
+    public int id { get; }
 
     int min;
     int max;
@@ -69,7 +69,7 @@ public class PlayInfo
         this.min = 0;
     }
 
-    public PlayInfo(PlayInfo origin, ulong id)
+    public PlayInfo(PlayInfo origin, int id)
     {
         this.id = id;
         this.name = origin.name;
